Move risk budget settings to Step 1
