+++ src/components/calculator/ExplanationCard/CalculationBreakdown.scss
