+++ src/locales/en.json
@@ -72,7 +72,7 @@
       "risk_budget_explanation": "You selected an annual budget of a {{ budget_percent }} chance of getting COVID per year ({{ budget_points }} microCOVIDs). That equates to a weekly budget of <strong>{{ budget_weekly }} microCOVIDs per week</strong>.",
       "risk_budget_used": "Budget used",
       "risk_budget_usage_explanation": "This interaction uses <strong>{{ usage }}</strong>. This amount of your budget will be used <strong>{{ frequency }}</strong>.",
-      "risk_tolerance_selector_header": "Adjust your risk tolerance",
+      "risk_tolerance_selector_header": "Your risk tolerance",
       "last_updated": "Model last updated on {{ date }}. <1>See all updates…</1>"
     },
     "pointsdisplay": {
@@ -114,7 +114,7 @@
     "prevalence": {
       "reported_prevalence": "Reported prevalence",
       "adjusted_prevalence": "Adjusted prevalence",
-      "details_header": "Details",
+      "details_header": "Location details",
       "last_week_cases": "Reported cases in past week",
       "population": "Total population",
       "cases_stable_or_decreasing": "Cases are stable or decreasing.",
@@ -123,7 +123,7 @@
       "data_last_updated": "Data last updated",
       "completed_vaccinations": "Percent of population fully vaccinated"
     },
-    "location_selector_header": "Step 1: Enter your location",
+    "step_1_header": "Step 1: Your settings",
     "location_subprompt": {
       "country_or_regions": "Entire country, or select region…",
       "CA": "Entire country, or select province/territory…",
