+++ src/locales/fr.json
@@ -100,7 +100,7 @@
         "prevalence_info_source_information": "Données de prévalence consolidées de <1>Johns Hopkins CSSE</1> (cas signalés), <4>Covid Act Now</4> (taux de tests positifs aux États-Unis) et <7>Our World in Data</7> (taux internationaux de tests positifs).",
         "select_location_enter_manually": "Entrez les données manuellement…",
         "select_location_placeholder": "Sélectionnez le pays ou l'État américain...",
-        "location_selector_header": "Étape 1 : Entrez votre localisation",
+        "step_1_header": "Étape 1 : Entrez votre localisation",
         "location_subprompt_US": "État entier ou sélectionnez un comté…",
         "location_subprompt_US-AK": "État entier ou sélectionnez un arrondissement…",
         "location_subprompt_US-LA": "État entier ou sélectionnez une paroisse…",
